Move context files to own folder
