+++ frontend/src/Components/PermissionRoute/PermissionRoute.tsx
@@ -1,4 +1,4 @@
-import { useAuthContext } from '~/AuthContext';
+import { useAuthContext } from '~/context/AuthContext';
 import { AccessDenied } from '~/Components';
 import { Children } from '~/types';
 import { hasPerm } from '~/utils';
