+++ frontend/src/Components/SamfundetLogoSpinner/SamfundetLogoSpinner.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
-import { AuthContextProvider } from '~/AuthContext';
-import { GlobalContextProvider } from '~/GlobalContextProvider';
+import { AuthContextProvider } from '~/context/AuthContext';
+import { GlobalContextProvider } from '~/context/GlobalContextProvider';
 import { SamfundetLogoSpinner } from './SamfundetLogoSpinner';
 
 // Local component config.
