+++ frontend/src/context/GlobalContextProvider.tsx
@@ -1,11 +1,11 @@
 import axios from 'axios';
 import { createContext, useContext, useEffect, useState } from 'react';
-import { useAuthContext } from '~/AuthContext';
+import { useAuthContext } from './AuthContext';
 import { getAllNotifications, getCsrfToken, getKeyValues, putUserPreference } from '~/api';
 import { MIRROR_CLASS, MOBILE_NAVIGATION_OPEN, ThemeValue, XCSRFTOKEN } from '~/constants';
 import { useMouseTrail, useTheme } from '~/hooks';
 import { Children, KeyValueMap, SetState } from '~/types';
-import { NotificationDto } from './dto';
+import { NotificationDto } from '~/dto';
 
 /**
  * Define which values the global context can contain.
