+++ frontend/src/hooks.ts
@@ -1,13 +1,13 @@
 import { MutableRefObject, useEffect, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useLocation, useNavigate } from 'react-router-dom';
-import { useAuthContext } from '~/AuthContext';
-import { useGlobalContext } from '~/GlobalContextProvider';
+import { useAuthContext } from './context/AuthContext';
+import { useGlobalContext } from './context/GlobalContextProvider';
 import { getTextItem, putUserPreference } from '~/api';
 import { Key, SetState } from '~/types';
 import { createDot, hasPerm, isTruthy, updateBodyThemeClass } from '~/utils';
 import { LinkTarget } from './Components/Link/Link';
-import { BACKEND_DOMAIN, THEME, THEME_KEY, ThemeValue, desktopBpLower, mobileBpUpper } from './constants';
+import { BACKEND_DOMAIN, desktopBpLower, mobileBpUpper, THEME, THEME_KEY, ThemeValue } from './constants';
 import { TextItemDto } from './dto';
 import { LANGUAGES } from './i18n/constants';
