+++ frontend/src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { AuthContextProvider } from '~/AuthContext';
+import { AuthContextProvider } from '~/context/AuthContext';
 import '~/global.scss';
-import { GlobalContextProvider } from '~/GlobalContextProvider';
+import { GlobalContextProvider } from '~/context/GlobalContextProvider';
 import { reportWebVitals } from '~/reportWebVitals';
 import { RouterProvider } from 'react-router-dom';
 import { router } from '~/AppRoutes';
